+++ app/javascript/packs/components/Page.vue
@@ -34,8 +34,11 @@ export default Vue.extend({
       this.editing = false;
     },
     destroy() {
-      this.$store.commit("setPageId", this.page.id);
-      this.$store.dispatch("destroyPage");
+      const confirmed = window.confirm("Are you sure ?")
+      if (confirmed) {
+        this.$store.commit("setPageId", this.page.id);
+        this.$store.dispatch("destroyPage");
+      }
     }
   }
 });
